Extract compose_chain_contracts into libprovekit per CCP
The Contract Composition Protocol (CCP) v1.0.0 specified at
protocol/specs/2026-05-09-contract-composition-protocol.md mandates
that compose_chain_contracts be a single canonical primitive in
libprovekit, callable identically from every language lifter and
verifier. CCP sections 2, 5, and 9 name the function, its
determinism guarantees, and the eight algebra rules; section 5
explicitly places the implementation at libprovekit/src/compose.rs.

Until now the function lived inside provekit-walk/src/contract.rs,
which meant only the Rust walking lifter could call it. The C
lifter, future Java / Go / TypeScript / Python lifters, and the
provekit-cli compose subcommand could not reach the canonical
implementation. Federation across languages depends on
byte-identical composed CIDs from a single source of truth.

This commit is plumbing only: move plus rewire plus smoke test.
The composition algebra is unchanged; CIDs are byte-identical
pre-refactor and post-refactor (the existing 238 walk lib tests
continue to pass without modification, including the deterministic
CID assertions that would catch any algebra drift).

Moved to libprovekit::compose:

  - compose_chain_contracts, ChainStep
  - compose_function_contracts, compose_function_contracts_checked,
    compose_with_composed
  - FunctionContractMemento, ComposedFunctionContract
  - EffectSet, Effect, AtomicKind
  - AliasingMemento, AliasingStatus
  - OpacityError, OpacityMementoLookup, EmptyOpacityPool,
    PinInvariantMementoView
  - Locus (data type only; the syn-driven from_span constructor
    stays in walk so libprovekit doesn't gain a syn dependency)
  - build_value, build_memento_value, sort_to_value
  - The capture-avoiding substitute_in_formula and the JCS / CID
    canonical-encoding glue (formula_to_canonical, cid_of_value,
    jcs_bytes_of_value), duplicated from walk's wp.rs and
    canonical.rs so the compose module is self-contained. Walk's
    pre-existing copies remain untouched; both impls operate over
    identical types from provekit-ir-types and produce
    byte-equivalent output by construction.

Walk's contract.rs is now a thin shim: pub use everything from
libprovekit::compose, plus the syn-walking helpers
(build_function_contract, detect_effects, scan_*_for_effects,
extract_formals, infer_sort, is_io_method, is_known_pure_method,
is_known_pure_call) and the existing 32 tests. Existing
'use crate::contract::FunctionContractMemento' paths in walk's
other modules (chain.rs, marriage.rs, llbc_calls.rs, llbc_lift.rs,
envelope.rs, type_decl.rs, charon_runner.rs) all continue to
resolve through the re-export, so no churn outside locus.rs and
contract.rs.

Walk's locus.rs adopts libprovekit::compose::Locus via re-export
and supplies a syn-using from_span free function plus a
LocusFromSpanExt trait so the existing five Locus::from_span
call sites in type_decl.rs keep working with one extra import.

provekit-verifier and provekit-cli only refer to the compose
primitive in comments today, so no rewiring was needed there.
They continue to build clean.

Smoke test at libprovekit/tests/compose_smoke.rs composes two
trivial pure atoms via compose_chain_contracts and asserts the
composed CID is byte-stable across runs against a pinned hex
value. Per CCP section 5 ("Test corpus") and section 11
("Versioning and revocation"), any future change that alters this
pinned CID requires a CCP version bump; the test is the
conformance witness for v1.0.0 of the algebra.

No FFI work; the C ABI wrapper at libprovekit/include/provekit-compose.h
and the JSON-RPC subprocess CLI subcommand are follow-up commits
(spec sections 6.2 and 6.3, both marked planned in Appendix B).
The C lifter's federation hookup is a separate agent's PR in
parallel.

Build verified:
  cargo build --release -p libprovekit -p provekit-walk -p provekit-cli
  cargo test -p libprovekit                 # 1 passed (compose_smoke)
  cargo test -p provekit-walk --lib         # 238 passed
  cargo test -p provekit-walk               # 238 + 12 integration files passed

Pinned smoke-test CID for the two-atom identity-result fixture:
blake3-512:36212b7bf7b9ccf264950940a33d64e1cfe88b6f4d8a47c01949fc64d9359d1813d6147aa2e1afe82b01e6e7ebcbe0a413683284b5f47ffef5bf364213304665
