+++ implementations/rust/libprovekit/Cargo.toml
@@ -8,6 +8,7 @@ description = "ProvekIt Rust reference library for protocol workflows."
 
 [dependencies]
 provekit-canonicalizer = { path = "../provekit-canonicalizer" }
+provekit-ir-types = { path = "../provekit-ir-types" }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
