+++ implementations/rust/provekit-walk/Cargo.toml
@@ -30,6 +30,7 @@ provekit-ir-types = { path = "../provekit-ir-types" }
 provekit-canonicalizer = { path = "../provekit-canonicalizer" }
 provekit-claim-envelope = { path = "../provekit-claim-envelope" }
 provekit-proof-envelope = { path = "../provekit-proof-envelope" }
+libprovekit = { path = "../libprovekit" }
 serde_json = { workspace = true }
 base64 = { workspace = true }
 thiserror = { workspace = true }
