+++ implementations/rust/provekit-walk/src/contract.rs
@@ -1,436 +1,38 @@
 // SPDX-License-Identifier: Apache-2.0
 //
-// FunctionContractMemento: every function's externally-visible
-// behavior collapsed to a content-addressed memento. Per #376.
+// Walk-side AST builders and effect detector for FunctionContractMemento.
 //
-// Each function in the lifted source emits ONE FunctionContractMemento.
-// Pure functions' contracts compose by hash combination, collapsing
-// arbitrarily-deep call subtrees into single CIDs (paper 07 §6's
-// "compose for free, compress to nothing"). Impure functions also
-// emit contract mementos for cross-program lookup but cannot compose.
+// The composition algebra (compose_chain_contracts, compose_function_contracts,
+// compose_with_composed, compose_function_contracts_checked, the supporting
+// data types FunctionContractMemento / ComposedFunctionContract / EffectSet /
+// Effect / OpacityError / OpacityMementoLookup / ChainStep / etc.) lives at
+// the workspace level in `libprovekit::compose` per the Contract Composition
+// Protocol (CCP) spec, sections 2 / 5 / 9
+// (protocol/specs/2026-05-09-contract-composition-protocol.md).
 //
-// Per Sir's design constraints:
-//   1. Singular formal substitution: each formal arrival is its own
-//      composition. f(a, g(b), c) yields three arrivals; only the
-//      g-substitution composes f's contract with g's contract.
-//   2. CID-namespaced result variable: each contract's post uses
-//      `result` locally, but composition renames inner.result to
-//      result_<inner.cid> before substituting into outer.pre, so
-//      different functions' results never collide.
-//   3. Effect-set from the start, not a one-bit pure marker. Pure is
-//      the empty set. Adding effect variants over time is content-
-//      addressing-safe (existing CIDs unchanged).
-//   4. Every function emits; composition refuses non-pure.
-//
-// Schema (canonical bytes, JCS-encoded):
-//
-//   FunctionContractMemento:
-//   {
-//     "schemaVersion": "1",
-//     "kind": "function-contract",
-//     "fnName": <string>,
-//     "formals": [<string>, ...],          // parameter names
-//     "formalSorts": [<Sort>, ...],         // per-formal sort
-//     "returnSort": <Sort>,
-//     "pre": <IrFormula>,
-//     "post": <IrFormula>,                  // references `result`
-//     "bodyCid": <CID or null>,             // shadow-source CID of body
-//     "effects": [<Effect>, ...]            // empty = pure
-//   }
-//
-//   Effect:
-//   { "kind": "reads", "target": <string> }
-//   { "kind": "writes", "target": <string> }
-//   { "kind": "io" }
-//   { "kind": "unsafe" }
-//   { "kind": "panics" }
-//   { "kind": "unresolved_call", "name": <string> }
+// This module re-exports those types and the canonical-encoding glue (so
+// existing `use crate::contract::*` paths in walk's other modules continue
+// to resolve) and adds the syn-driven helpers that BUILD a FunctionContract
+// Memento from a `syn::ItemFn`. AST traversal stays in walk; the algebra
+// lives once, in libprovekit.
 
-use std::sync::Arc;
-
-use provekit_canonicalizer::Value;
-use provekit_ir_types::{IrFormula, IrTerm, Sort};
 use syn::{Expr, ExprUnsafe, FnArg, ItemFn, Pat, Stmt};
 
-use crate::canonical::{cid_of_value, formula_to_canonical, jcs_bytes_of_value};
-use crate::lift::{lift_function_postcondition, lift_function_precondition};
-use crate::locus::Locus;
-use crate::wp::{substitute_in_formula, Wp};
+// ---- Re-export the canonical algebra and supporting types ----
 
-// ---- Effect set ----
+pub use libprovekit::compose::{
+    build_memento_value, build_value, cid_of_value, compose_chain_contracts,
+    compose_function_contracts, compose_function_contracts_checked, compose_with_composed,
+    formula_to_canonical, jcs_bytes_of_value, sort_to_value, substitute_in_formula,
+    AliasingMemento, AliasingStatus, AtomicKind, ChainStep, ComposedFunctionContract, Effect,
+    EffectSet, EmptyOpacityPool, FunctionContractMemento, Locus, OpacityError,
+    OpacityMementoLookup, PinInvariantMementoView,
+};
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Effect {
-    /// Reads a named state cell (global, capture, mut binding).
-    Reads { target: String },
-    /// Writes a named state cell.
-    Writes { target: String },
-    /// Performs IO (println!, file I/O, network, syscall).
-    Io,
-    /// Contains an unsafe block.
-    Unsafe,
-    /// May panic on inputs satisfying the precondition.
-    Panics,
-    /// Calls a function whose effect-set is unknown to the lifter.
-    UnresolvedCall { name: String },
-    /// Contains a loop whose invariant has not been supplied. The
-    /// `loop_cid` is the JCS-byte BLAKE3-512 hash of the loop's
-    /// LLBC body block — it identifies the loop independent of the
-    /// containing function. A separate `LoopInvariantMemento`
-    /// keyed by this loop_cid supplies the invariant + decreasing
-    /// function; the substrate refuses to compose this contract
-    /// downstream until that memento is present and verified.
-    /// Per paper 07 §11 — loops are first-class deferred memento
-    /// targets.
-    OpaqueLoop { loop_cid: String },
-    /// Contains a `?` operator (Try-branch with early return). The
-    /// `try_cid` is the content hash of the Switch::Match block that
-    /// implements the Try-branch shape. A separate
-    /// `TryBranchMemento` (or specific Result/Option-shape spec)
-    /// supplies the success-path/failure-path contract pair. The
-    /// substrate refuses composition until that memento lands —
-    /// honest opacity rather than silent assumption that all `?`
-    /// callers handle the Err branch. Same shape of opacity as
-    /// OpaqueLoop.
-    EarlyReturn { try_cid: String },
-    /// Constructs a closure value. The closure body is itself a
-    /// regular fun_decl (Charon emits it as a Fn/FnMut/FnOnce trait
-    /// impl method) — its contract is lifted normally through the
-    /// usual pipeline. This effect records the link between the
-    /// CAPTURING site (this function) and the body fn: when the
-    /// closure is later called, the call composes against the body
-    /// contract. The substrate uses this effect to recognize that
-    /// composition through the call is mediated by the captured
-    /// environment. `body_fn_cid` is the body fun_decl's content_cid
-    /// once it's lifted; `n_captures` is the count of captured
-    /// values bundled at this site.
-    ClosureCapture {
-        body_fn_cid: String,
-        n_captures: usize,
-    },
-    /// Accepts a `Pin<P>` formal parameter. The substrate treats
-    /// Pin'd references as opaque until a PinProjectionMemento
-    /// supplies the projection safety proof. `target` is the formal
-    /// parameter name.
-    PinnedReference { target: String },
-    /// Accepts or dereferences a raw pointer (`*const T` or `*mut T`)
-    /// in a formal parameter position. `target` is the formal
-    /// parameter name; `mutable` distinguishes `*mut` from `*const`.
-    /// This is always co-present with `Effect::Unsafe` since raw
-    /// pointer derefs require an unsafe block to be safe-Rust
-    /// observable, but the pin/rawptr detection fires on the formal
-    /// type, not on a dereference site — so `Unsafe` is emitted
-    /// independently by the unsafe-block scanner.
-    RawPointerProvenance { target: String, mutable: bool },
-    /// Calls an atomic intrinsic (core::sync::atomic::Atomic*).
-    /// `target` is the local variable (or expression) being accessed;
-    /// `kind` is the operation class; `ordering` is the memory-order
-    /// string when it can be statically determined (may be None when
-    /// the ordering is passed as a runtime argument).
-    AtomicAccess {
-        target: String,
-        kind: AtomicKind,
-        ordering: Option<String>,
-    },
-    /// Emitted when a function has formal parameters that are shared
-    /// references (&T) to types with interior mutability. The substrate
-    /// refuses composition unless every pair in `formals` has a matching
-    /// AliasingMemento in the verifier pool.
-    PossibleAliasing {
-        /// Sorted lexicographically for JCS byte-determinism.
-        formals: Vec<String>,
-    },
-    /// Calls `drop_in_place`. Drops can run user-defined `Drop::drop`,
-    /// allocate, panic, or perform arbitrary side effects. The
-    /// substrate refuses to compose a contract carrying this effect
-    /// without an explicit discharge memento (or the callee's own
-    /// lifted contract, once the drop function body is liftable).
-    /// `name` is the name of the type being dropped.
-    Drop { name: String },
-}
-
-/// Operation class for `Effect::AtomicAccess`.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AtomicKind {
-    /// Atomic load (load, peek).
-    Load,
-    /// Atomic store.
-    Store,
-    /// Read-modify-write (fetch_add, fetch_sub, fetch_and, fetch_or,
-    /// fetch_xor, fetch_nand, swap, fetch_update).
-    Rmw,
-    /// Compare-and-swap (compare_exchange, compare_exchange_weak,
-    /// compare_and_swap [deprecated]).
-    Cas,
-}
-
-impl AtomicKind {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            AtomicKind::Load => "load",
-            AtomicKind::Store => "store",
-            AtomicKind::Rmw => "rmw",
-            AtomicKind::Cas => "cas",
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct AliasingMemento {
-    pub formal_a: String,
-    pub formal_b: String,
-    pub status: AliasingStatus,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AliasingStatus {
-    Disjoint,
-    MaybeAlias,
-}
-
-impl AliasingStatus {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            AliasingStatus::Disjoint => "Disjoint",
-            AliasingStatus::MaybeAlias => "MaybeAlias",
-        }
-    }
-}
-
-impl AliasingMemento {
-    pub fn to_jcs_value(&self) -> Arc<Value> {
-        debug_assert!(
-            self.formal_a <= self.formal_b,
-            "AliasingMemento invariants violated: formal_a ({}) must be <= formal_b ({})",
-            self.formal_a,
-            self.formal_b
-        );
-        Value::object([
-            ("kind", Value::string("aliasing-memento")),
-            ("formal_a", Value::string(self.formal_a.clone())),
-            ("formal_b", Value::string(self.formal_b.clone())),
-            ("status", Value::string(self.status.as_str())),
-        ])
-    }
-}
-
-impl Effect {
-    fn to_value(&self) -> Arc<Value> {
-        match self {
-            Effect::Reads { target } => Value::object([
-                ("kind", Value::string("reads")),
-                ("target", Value::string(target.clone())),
-            ]),
-            Effect::Writes { target } => Value::object([
-                ("kind", Value::string("writes")),
-                ("target", Value::string(target.clone())),
-            ]),
-            Effect::Io => Value::object([("kind", Value::string("io"))]),
-            Effect::Unsafe => Value::object([("kind", Value::string("unsafe"))]),
-            Effect::Panics => Value::object([("kind", Value::string("panics"))]),
-            Effect::UnresolvedCall { name } => Value::object([
-                ("kind", Value::string("unresolved_call")),
-                ("name", Value::string(name.clone())),
-            ]),
-            Effect::OpaqueLoop { loop_cid } => Value::object([
-                ("kind", Value::string("opaque_loop")),
-                ("loopCid", Value::string(loop_cid.clone())),
-            ]),
-            Effect::EarlyReturn { try_cid } => Value::object([
-                ("kind", Value::string("early_return")),
-                ("tryCid", Value::string(try_cid.clone())),
-            ]),
-            Effect::ClosureCapture {
-                body_fn_cid,
-                n_captures,
-            } => Value::object([
-                ("kind", Value::string("closure_capture")),
-                ("bodyFnCid", Value::string(body_fn_cid.clone())),
-                ("nCaptures", Value::integer(*n_captures as i64)),
-            ]),
-            Effect::PinnedReference { target } => Value::object([
-                ("kind", Value::string("pinned_reference")),
-                ("target", Value::string(target.clone())),
-            ]),
-            Effect::RawPointerProvenance { target, mutable } => Value::object([
-                ("kind", Value::string("raw_ptr_provenance")),
-                ("mutable", Value::boolean(*mutable)),
-                ("target", Value::string(target.clone())),
-            ]),
-            Effect::AtomicAccess {
-                target,
-                kind,
-                ordering,
-            } => {
-                let mut fields: Vec<(&str, Arc<Value>)> = vec![
-                    ("kind", Value::string("atomic_access")),
-                    ("atomicKind", Value::string(kind.as_str())),
-                    ("target", Value::string(target.clone())),
-                ];
-                if let Some(ord) = ordering {
-                    fields.push(("ordering", Value::string(ord.clone())));
-                }
-                Value::object(fields)
-            }
-            Effect::PossibleAliasing { formals } => {
-                let formals_arr: Vec<Arc<Value>> =
-                    formals.iter().map(|f| Value::string(f.clone())).collect();
-                Value::object([
-                    ("kind", Value::string("possible_aliasing")),
-                    ("formals", Value::array(formals_arr)),
-                ])
-            }
-            Effect::Drop { name } => Value::object([
-                ("kind", Value::string("drop")),
-                ("name", Value::string(name.clone())),
-            ]),
-        }
-    }
-
-    fn sort_key(&self) -> String {
-        // Stable string for sorting effects in the canonical encoding.
-        match self {
-            Effect::Reads { target } => format!("0:reads:{}", target),
-            Effect::Writes { target } => format!("1:writes:{}", target),
-            Effect::Io => "2:io".to_string(),
-            Effect::Unsafe => "3:unsafe".to_string(),
-            Effect::Panics => "4:panics".to_string(),
-            Effect::UnresolvedCall { name } => format!("5:unresolved:{}", name),
-            Effect::OpaqueLoop { loop_cid } => format!("6:opaque_loop:{}", loop_cid),
-            Effect::EarlyReturn { try_cid } => format!("7:early_return:{}", try_cid),
-            Effect::ClosureCapture {
-                body_fn_cid,
-                n_captures,
-            } => format!("8:closure_capture:{}:{}", body_fn_cid, n_captures),
-            Effect::PinnedReference { target } => format!("9:pinned_reference:{}", target),
-            Effect::RawPointerProvenance { target, mutable } => {
-                format!("10:raw_ptr_provenance:{}:{}", target, mutable)
-            }
-            Effect::AtomicAccess {
-                target,
-                kind,
-                ordering,
-            } => format!(
-                "11:atomic_access:{}:{}:{}",
-                target,
-                kind.as_str(),
-                ordering.as_deref().unwrap_or("")
-            ),
-            Effect::PossibleAliasing { formals } => {
-                format!("13:possible_aliasing:{}", formals.join(","))
-            }
-            Effect::Drop { name } => format!("14:drop:{}", name),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct EffectSet {
-    pub effects: Vec<Effect>,
-}
-
-impl EffectSet {
-    pub fn empty() -> Self {
-        Self { effects: vec![] }
-    }
-    pub fn is_pure(&self) -> bool {
-        self.effects.is_empty()
-    }
-    pub fn add(&mut self, e: Effect) {
-        if !self.effects.iter().any(|x| x == &e) {
-            self.effects.push(e);
-        }
-    }
-    fn to_value(&self) -> Arc<Value> {
-        let mut sorted = self.effects.clone();
-        sorted.sort_by(|a, b| a.sort_key().cmp(&b.sort_key()));
-        let items: Vec<Arc<Value>> = sorted.iter().map(|e| e.to_value()).collect();
-        Value::array(items)
-    }
-}
-
-// ---- Contract memento ----
-
-#[derive(Debug, Clone)]
-pub struct FunctionContractMemento {
-    pub fn_name: String,
-    pub formals: Vec<String>,
-    pub formal_sorts: Vec<Sort>,
-    pub formal_regions: Vec<Option<String>>,
-    pub return_sort: Sort,
-    pub return_region: Option<String>,
-    pub pre: IrFormula,
-    pub post: IrFormula,
-    pub body_cid: Option<String>,
-    pub effects: EffectSet,
-    pub locus: Locus,
-    pub canonical_bytes: Vec<u8>,
-    pub cid: String,
-    pub auto_minted_mementos: Vec<AliasingMemento>,
-}
-
-impl FunctionContractMemento {
-    pub fn is_pure(&self) -> bool {
-        self.effects.is_pure()
-    }
-
-    /// Extract the term-side of `result = <expr>` from the post, used
-    /// when composing this contract's result into another contract's
-    /// pre/post. Returns None if the post doesn't have a recognizable
-    /// `result = ...` equation.
-    pub fn result_value(&self) -> Option<IrTerm> {
-        find_result_equation(&self.post, "result")
-    }
-
-    /// Short tag used to namespace this contract's `result` variable
-    /// when composing. Uses the CID's hex tail.
-    pub fn result_var_name(&self) -> String {
-        let tail: String = self
-            .cid
-            .chars()
-            .rev()
-            .take(12)
-            .collect::<Vec<_>>()
-            .into_iter()
-            .rev()
-            .collect();
-        format!("result__{}", tail)
-    }
-
-    /// Check that every `Effect::PossibleAliasing` pair of formals has a
-    /// matching AliasingMemento in `auto_minted_mementos`. Returns
-    /// `Ok(())` if aliasing is fully discharged or no PossibleAliasing
-    /// effects are present. Returns the first undischarged pair as an error.
-    pub fn check_aliasing_discharged(&self) -> Result<(), OpacityError> {
-        for effect in &self.effects.effects {
-            if let Effect::PossibleAliasing { formals } = effect {
-                if formals.len() < 2 {
-                    continue;
-                }
-                for i in 0..formals.len() {
-                    for j in (i + 1)..formals.len() {
-                        let a = &formals[i];
-                        let b = &formals[j];
-                        let covered = self.auto_minted_mementos.iter().any(|m| {
-                            (m.formal_a == *a && m.formal_b == *b)
-                                || (m.formal_a == *b && m.formal_b == *a)
-                        });
-                        if !covered {
-                            return Err(OpacityError::AliasingNotDischarged {
-                                formal_a: a.clone(),
-                                formal_b: b.clone(),
-                            });
-                        }
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-}
+// ---- AST builders ----
 
 /// Build a FunctionContractMemento for an `ItemFn`. The body_cid is
-/// optional — pass None when the body's shadow source isn't computed
+/// optional; pass None when the body's shadow source isn't computed
 /// (e.g., during a lift-only pass). The locus carries source-position
 /// metadata for downstream developer feedback; pass None to use an
 /// unknown/empty locus.
@@ -451,10 +53,10 @@ pub fn build_function_contract_with_file(
     let fn_name = item_fn.sig.ident.to_string();
     let (formals, formal_sorts) = extract_formals(item_fn);
     let return_sort = extract_return_sort(item_fn);
-    let pre = lift_function_precondition(item_fn).into_formula();
-    let post = lift_function_postcondition(item_fn).into_formula();
+    let pre = crate::lift::lift_function_precondition(item_fn).into_formula();
+    let post = crate::lift::lift_function_postcondition(item_fn).into_formula();
     let effects = detect_effects(item_fn);
-    let locus = Locus::from_span(item_fn.sig.ident.span(), file_path);
+    let locus = crate::locus::from_span(item_fn.sig.ident.span(), file_path);
 
     let value = build_value(
         &fn_name,
@@ -486,677 +88,6 @@ pub fn build_function_contract_with_file(
         canonical_bytes,
         cid,
         auto_minted_mementos: vec![],
-    }
-}
-
-fn build_value(
-    fn_name: &str,
-    formals: &[String],
-    formal_sorts: &[Sort],
-    return_sort: &Sort,
-    pre: &IrFormula,
-    post: &IrFormula,
-    body_cid: Option<&str>,
-    effects: &EffectSet,
-    locus: &Locus,
-    auto_minted_mementos: &[AliasingMemento],
-) -> Arc<Value> {
-    let formals_arr: Vec<Arc<Value>> = formals.iter().map(|n| Value::string(n.clone())).collect();
-    let formal_sorts_arr: Vec<Arc<Value>> = formal_sorts.iter().map(|s| sort_to_value(s)).collect();
-    let body_cid_val: Arc<Value> = match body_cid {
-        Some(c) => Value::string(c.to_string()),
-        None => Value::null(),
-    };
-    Value::object([
-        ("schemaVersion", Value::string("1")),
-        ("kind", Value::string("function-contract")),
-        ("fnName", Value::string(fn_name.to_string())),
-        ("formals", Value::array(formals_arr)),
-        ("formalSorts", Value::array(formal_sorts_arr)),
-        ("returnSort", sort_to_value(return_sort)),
-        ("pre", formula_to_canonical(pre)),
-        ("post", formula_to_canonical(post)),
-        ("bodyCid", body_cid_val),
-        ("effects", effects.to_value()),
-        ("locus", locus.to_value()),
-        (
-            "autoMintedMementos",
-            aliasing_mementos_to_value(auto_minted_mementos),
-        ),
-    ])
-}
-
-/// Build the canonical `Value` for a `FunctionContractMemento`. Used
-/// by callers (LLBC lift, marriage) that need to recompute the
-/// memento's canonical bytes / CID after replacing fields like the
-/// pre/post formulas.
-pub fn build_memento_value(c: &FunctionContractMemento) -> Arc<Value> {
-    build_value(
-        &c.fn_name,
-        &c.formals,
-        &c.formal_sorts,
-        &c.return_sort,
-        &c.pre,
-        &c.post,
-        c.body_cid.as_deref(),
-        &c.effects,
-        &c.locus,
-        &c.auto_minted_mementos,
-    )
-}
-
-/// Canonical JCS array for a set of AliasingMementos. Entries are sorted
-/// lexicographically by (formal_a, formal_b) for byte-determinism. Empty
-/// list is always included as `[]` (never omitted) so the CID is
-/// consistent across deployments.
-fn aliasing_mementos_to_value(mementos: &[AliasingMemento]) -> Arc<Value> {
-    let mut sorted: Vec<&AliasingMemento> = mementos.iter().collect();
-    sorted.sort_by(|a, b| {
-        a.formal_a
-            .cmp(&b.formal_a)
-            .then_with(|| a.formal_b.cmp(&b.formal_b))
-    });
-    let items: Vec<Arc<Value>> = sorted.iter().map(|m| m.to_jcs_value()).collect();
-    Value::array(items)
-}
-
-fn sort_to_value(s: &Sort) -> Arc<Value> {
-    match s {
-        Sort::Primitive { name } => Value::object([
-            ("kind", Value::string("primitive")),
-            ("name", Value::string(name.clone())),
-        ]),
-        // Function / Dependent sorts (added by #361 for the
-        // dependent-type / first-class-function-sort grammar
-        // expansion) aren't currently produced by the AST/LLBC
-        // walk's `synth_item_fn` shell — we always emit a plain
-        // primitive sort. Emit them as opaque so the contract
-        // canonical bytes stay valid if a downstream caller passes
-        // them through. Translating them faithfully into our IR
-        // sort encoding is tracked as part of #384 Tier A.1
-        // (type-aware predicate sorts).
-        Sort::Function { .. } | Sort::Dependent { .. } => Value::object([
-            ("kind", Value::string("opaque")),
-            (
-                "reason",
-                Value::string("function-or-dependent-sort-not-yet-modeled"),
-            ),
-        ]),
-        // IEEE-754 float sort (added by #385). Canonical encoding carries
-        // `kind: "float"` and the bit-width so downstream solvers can pick
-        // the right float theory. NaN/inf/ordering semantics are deliberately
-        // NOT modelled at this layer; see Sort::Float doc comment.
-        Sort::Float { width } => Value::object([
-            ("kind", Value::string("float")),
-            ("width", Value::integer(i64::from(*width))),
-        ]),
-        // RegionSort (added by #401). Canonical encoding carries kind + name
-        // so contract CIDs for lifetime-parameterised functions are
-        // distinguishable from primitives and CID-stable across renames.
-        Sort::Region { name } => Value::object([
-            ("kind", Value::string("region")),
-            ("name", Value::string(name.clone())),
-        ]),
-    }
-}
-
-// ---- Opacity discharge ----
-
-/// Trait for querying whether a discharge memento is present for a
-/// given opacity effect. Implemented by `MementoPool` in
-/// `provekit-verifier`; in-crate tests use a mock.
-///
-/// The `has_*` methods return `true` iff a conforming discharge memento
-/// for that specific site is present in the pool and has passed all
-/// validation rules defined in the corresponding spec. The
-/// `lookup_pin_invariant` method returns `Option<PinInvariantMementoView>`
-/// to carry the discharge-relevant fields (function_cid, pinned_target,
-/// invariant) for downstream non-emptiness validation.
-///
-/// Specs:
-///   - `LoopInvariantMemento`  (protocol/specs/2026-05-05-loop-invariant-memento.md)
-///   - `TryBranchMemento`      (protocol/specs/2026-05-05-try-branch-memento.md)
-///   - `ClosureBindingMemento` (protocol/specs/2026-05-05-closure-binding-memento.md)
-///   - `PinInvariantMemento`   (protocol/specs/2026-05-05-pin-invariant-memento.md)
-///   - `UnresolvedCall`:       no memento kind yet; always undischarged.
-pub trait OpacityMementoLookup {
-    fn has_loop_invariant(&self, loop_cid: &str) -> bool;
-    fn has_try_branch(&self, try_cid: &str) -> bool;
-    fn has_closure_binding(&self, body_fn_cid: &str) -> bool;
-    /// Returns true when the pool contains a contract for the
-    /// type being dropped (i.e., the drop function has been lifted
-    /// and its effects have been reviewed). When false, composition
-    /// is refused — the substrate does not silently assume drops are
-    /// effect-free.
-    fn has_drop_contract(&self, type_name: &str) -> bool;
-    fn has_aliasing_memento(&self, formal_a: &str, formal_b: &str) -> bool;
-    fn lookup_pin_invariant(
-        &self,
-        function_cid: &str,
-        target: &str,
-    ) -> Option<PinInvariantMementoView>;
-}
-
-/// A no-op pool that never has any discharge mementos. Used as the
-/// default when callers don't yet track opacity discharge.
-pub struct EmptyOpacityPool;
-impl OpacityMementoLookup for EmptyOpacityPool {
-    fn has_loop_invariant(&self, _: &str) -> bool {
-        false
-    }
-    fn has_try_branch(&self, _: &str) -> bool {
-        false
-    }
-    fn has_closure_binding(&self, _: &str) -> bool {
-        false
-    }
-    fn has_drop_contract(&self, _: &str) -> bool {
-        false
-    }
-    fn has_aliasing_memento(&self, _: &str, _: &str) -> bool {
-        false
-    }
-    fn lookup_pin_invariant(
-        &self,
-        _function_cid: &str,
-        _target: &str,
-    ) -> Option<PinInvariantMementoView> {
-        None
-    }
-}
-
-/// Lightweight pool lookup type for PinInvariantMemento. The full
-/// memento has envelope/header/metadata layers (per the memento spec),
-/// but the pool only needs the discharge-relevant fields.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PinInvariantMementoView {
-    pub function_cid: String,
-    pub pinned_target: String,
-    pub invariant: String,
-}
-
-/// Error returned when composition is refused because an opacity effect
-/// is not discharged by a memento in the pool.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum OpacityError {
-    /// An `Effect::OpaqueLoop { loop_cid }` is present but no
-    /// `LoopInvariantMemento` with that `loopCid` is in the pool.
-    LoopNotDischarged { loop_cid: String },
-    /// An `Effect::EarlyReturn { try_cid }` is present but no
-    /// `TryBranchMemento` with that `tryCid` is in the pool.
-    EarlyReturnNotDischarged { try_cid: String },
-    /// An `Effect::ClosureCapture { body_fn_cid, .. }` is present but
-    /// no `ClosureBindingMemento` with that `bodyFnCid` is in the pool.
-    ClosureCaptureNotDischarged { body_fn_cid: String },
-    /// An `Effect::UnresolvedCall { name }` is present. No discharge
-    /// memento kind exists yet; composition is always refused.
-    UnresolvedCallNotDischarged { name: String },
-    /// An `Effect::PossibleAliasing` is present but the pair (formal_a, formal_b)
-    /// has no matching AliasingMemento in auto_minted_mementos.
-    AliasingNotDischarged { formal_a: String, formal_b: String },
-    /// An `Effect::Drop { name }` is present but no lifted drop
-    /// function contract for that type is in the pool.
-    DropNotDischarged { name: String },
-    /// An `Effect::PinnedReference { target }` is present but no
-    /// `PinInvariantMemento` with matching `pinnedTarget` is in the pool.
-    PinInvariantNotDischarged { target: String },
-}
-
-impl std::fmt::Display for OpacityError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::LoopNotDischarged { loop_cid } =>
-                write!(f, "opacity: OpaqueLoop(loopCid={loop_cid}) has no LoopInvariantMemento in pool"),
-            Self::EarlyReturnNotDischarged { try_cid } =>
-                write!(f, "opacity: EarlyReturn(tryCid={try_cid}) has no TryBranchMemento in pool"),
-            Self::ClosureCaptureNotDischarged { body_fn_cid } =>
-                write!(f, "opacity: ClosureCapture(bodyFnCid={body_fn_cid}) has no ClosureBindingMemento in pool"),
-            Self::UnresolvedCallNotDischarged { name } =>
-                write!(f, "opacity: UnresolvedCall({name}) has no discharge memento kind"),
-            Self::AliasingNotDischarged { formal_a, formal_b } =>
-                write!(f, "aliasing: PossibleAliasing pair ({formal_a}, {formal_b}) has no AliasingMemento in auto_minted_mementos"),
-            Self::DropNotDischarged { name } =>
-                write!(f, "opacity: Drop({name}) has no lifted drop function in pool"),
-            Self::PinInvariantNotDischarged { target } =>
-                write!(f, "opacity: PinnedReference(target={target}) has no PinInvariantMemento in pool"),
-        }
-    }
-}
-
-impl EffectSet {
-    /// Check whether all opacity effects in this set are discharged by
-    /// the given pool. Returns `Ok(())` iff:
-    ///   - there are no opacity effects at all, OR
-    ///   - every opacity effect has a corresponding memento in `pool`.
-    ///
-    /// The `function_cid` parameter is required for `PinnedReference`
-    /// discharge (which matches on `(function_cid, target)` pairs).
-    /// Pass `None` when the enclosing contract CID is not available
-    /// (in which case `PinnedReference` will always be undischarged).
-    ///
-    /// Non-opacity effects (Reads/Writes/Io/Unsafe/Panics) are NOT
-    /// checked here; composition with non-opacity effects is refused
-    /// by `compose_function_contracts` separately.
-    ///
-    /// Returns the FIRST undischarged opacity effect as an error.
-    /// The caller can collect all errors by iterating effects directly.
-    pub fn check_opacity_effects(
-        &self,
-        pool: &dyn OpacityMementoLookup,
-        function_cid: Option<&str>,
-    ) -> Result<(), OpacityError> {
-        for effect in &self.effects {
-            match effect {
-                Effect::OpaqueLoop { loop_cid } => {
-                    if !pool.has_loop_invariant(loop_cid) {
-                        return Err(OpacityError::LoopNotDischarged {
-                            loop_cid: loop_cid.clone(),
-                        });
-                    }
-                }
-                Effect::EarlyReturn { try_cid } => {
-                    if !pool.has_try_branch(try_cid) {
-                        return Err(OpacityError::EarlyReturnNotDischarged {
-                            try_cid: try_cid.clone(),
-                        });
-                    }
-                }
-                Effect::ClosureCapture { body_fn_cid, .. } => {
-                    if !pool.has_closure_binding(body_fn_cid) {
-                        return Err(OpacityError::ClosureCaptureNotDischarged {
-                            body_fn_cid: body_fn_cid.clone(),
-                        });
-                    }
-                }
-                Effect::UnresolvedCall { name } => {
-                    return Err(OpacityError::UnresolvedCallNotDischarged { name: name.clone() });
-                }
-                Effect::Drop { name } => {
-                    if !pool.has_drop_contract(name) {
-                        return Err(OpacityError::DropNotDischarged { name: name.clone() });
-                    }
-                }
-                // PossibleAliasing is discharged by auto_minted_mementos in the
-                // PossibleAliasing is discharged by auto_minted_mementos in the
-                // contract bundle; the substrate verifier checks these before
-                // composition. No pool check at this level.
-                Effect::PossibleAliasing { .. } => {}
-                Effect::PinnedReference { target } => {
-                    let fc = match function_cid {
-                        Some(cid) if !cid.is_empty() => cid,
-                        _ => {
-                            return Err(OpacityError::PinInvariantNotDischarged {
-                                target: target.clone(),
-                            })
-                        }
-                    };
-                    match pool.lookup_pin_invariant(fc, target) {
-                        Some(view) if !view.invariant.is_empty() => { /* discharged */ }
-                        _ => {
-                            return Err(OpacityError::PinInvariantNotDischarged {
-                                target: target.clone(),
-                            })
-                        }
-                    }
-                }
-                // Non-opacity effects: Reads/Writes/Io/Unsafe/Panics and the
-                // structural type-boundary effects (RawPointerProvenance,
-                // AtomicAccess) do not participate in memento-based discharge.
-                // They unconditionally block composition via the
-                // outer_non_opacity_pure check in compose_function_contracts_checked.
-                Effect::Reads { .. }
-                | Effect::Writes { .. }
-                | Effect::Io
-                | Effect::Unsafe
-                | Effect::Panics
-                | Effect::RawPointerProvenance { .. }
-                | Effect::AtomicAccess { .. } => {}
-            }
-        }
-        Ok(())
-    }
-}
-
-// ---- Composition ----
-
-#[derive(Debug, Clone)]
-pub struct ComposedFunctionContract {
-    pub component_cids: Vec<String>,
-    pub formal_idx: usize,
-    pub pre: IrFormula,
-    pub post: IrFormula,
-    pub canonical_bytes: Vec<u8>,
-    pub cid: String,
-}
-
-/// Compose two function contracts: the inner contract's result feeds
-/// the outer contract's `formal_idx`-th formal.
-///
-/// Refuses (returns None) if either contract is impure.
-pub fn compose_function_contracts(
-    outer: &FunctionContractMemento,
-    inner: &FunctionContractMemento,
-    formal_idx: usize,
-) -> Option<ComposedFunctionContract> {
-    if !outer.is_pure() || !inner.is_pure() {
-        return None;
-    }
-    if formal_idx >= outer.formals.len() {
-        return None;
-    }
-
-    // Step 1: rename inner's `result` to namespaced form (CID-prefixed).
-    let inner_result_name = inner.result_var_name();
-    let inner_post_renamed = substitute_in_formula(
-        inner.post.clone(),
-        "result",
-        &IrTerm::Var {
-            name: inner_result_name.clone(),
-        },
-    );
-
-    // Step 2: extract the term equated with the renamed result.
-    let inner_value = match find_result_equation(&inner_post_renamed, &inner_result_name) {
-        Some(t) => t,
-        None => return None,
-    };
-
-    // Step 3: substitute outer's formal with the inner's result-value.
-    let outer_formal = &outer.formals[formal_idx];
-    let outer_pre_substituted =
-        substitute_in_formula(outer.pre.clone(), outer_formal, &inner_value);
-    let outer_post_substituted =
-        substitute_in_formula(outer.post.clone(), outer_formal, &inner_value);
-
-    // Step 4: compose.
-    //   pre  = inner.pre ∧ (inner.post → outer.pre[inner_value/formal])
-    //   post = outer.post[inner_value/formal]
-    let pre = IrFormula::And {
-        operands: vec![
-            inner.pre.clone(),
-            IrFormula::Implies {
-                operands: vec![inner_post_renamed.clone(), outer_pre_substituted],
-            },
-        ],
-    };
-    let post = outer_post_substituted;
-
-    // Step 5: content-address.
-    let value = Value::object([
-        ("schemaVersion", Value::string("1")),
-        ("kind", Value::string("composed-function-contract")),
-        (
-            "components",
-            Value::array(vec![
-                Value::string(outer.cid.clone()),
-                Value::string(inner.cid.clone()),
-            ]),
-        ),
-        ("formalIdx", Value::integer(formal_idx as i64)),
-        ("pre", formula_to_canonical(&pre)),
-        ("post", formula_to_canonical(&post)),
-    ]);
-    let canonical_bytes = jcs_bytes_of_value(&value);
-    let cid = cid_of_value(&value);
-
-    Some(ComposedFunctionContract {
-        component_cids: vec![outer.cid.clone(), inner.cid.clone()],
-        formal_idx,
-        pre,
-        post,
-        canonical_bytes,
-        cid,
-    })
-}
-
-/// Compose two function contracts with opacity discharge checking.
-///
-/// This is the substrate-verifier-facing API. It distinguishes three
-/// refusal reasons:
-///
-/// 1. `Err(OpacityError::*)` — a contract carries an opacity effect
-///    (`OpaqueLoop`, `EarlyReturn`, `ClosureCapture`, `UnresolvedCall`,
-///    `Drop`, `PinnedReference`) with no corresponding discharge memento
-///    in `pool`. The caller knows exactly which effect is undischarged.
-///
-/// 2. `Ok(None)` — composition refused for a non-opacity reason:
-///    an unconditionally-blocking effect (Reads/Writes/Io/Unsafe/Panics)
-///    is present, `formal_idx` is out of bounds, or the inner post has
-///    no result equation.
-///
-/// 3. `Ok(Some(composed))` — all opacity effects are discharged by
-///    mementos in `pool`, no unconditionally-blocking effects are
-///    present, and composition succeeded.
-///
-/// Design note: unlike `compose_function_contracts`, this function
-/// allows contracts whose ONLY impure effects are opacity effects that
-/// are fully discharged. The existing `compose_function_contracts`
-/// (which refuses on ANY effect) remains unchanged for pre-discharge
-/// callers.
-pub fn compose_function_contracts_checked(
-    outer: &FunctionContractMemento,
-    inner: &FunctionContractMemento,
-    formal_idx: usize,
-    pool: &dyn OpacityMementoLookup,
-) -> Result<Option<ComposedFunctionContract>, OpacityError> {
-    // Phase 1: check opacity effects on both sides. Returns the first
-    // undischarged opacity effect as a typed error.
-    outer
-        .effects
-        .check_opacity_effects(pool, Some(&outer.cid))?;
-    inner
-        .effects
-        .check_opacity_effects(pool, Some(&inner.cid))?;
-
-    // Phase 1.5: check aliasing discharge on both sides. PossibleAliasing
-    // must have a matching AliasingMemento in auto_minted_mementos for
-    // every pair of formals. Returns the first undischarged pair as an error.
-    outer.check_aliasing_discharged()?;
-    inner.check_aliasing_discharged()?;
-
-    // Phase 2: check for unconditionally-blocking effects. After opacity,
-    // aliasing, and pin-invariant discharge, only Reads/Writes/Io/Unsafe/Panics
-    // and the remaining structural type-boundary effects can still block.
-    // A contract that had ONLY opacity effects — all now discharged — has
-    // no remaining blockers; one that still has non-opacity impure effects
-    // is still refused.
-    let outer_non_opacity_pure = outer.effects.effects.iter().all(|e| {
-        matches!(
-            e,
-            Effect::OpaqueLoop { .. }
-                | Effect::EarlyReturn { .. }
-                | Effect::ClosureCapture { .. }
-                | Effect::UnresolvedCall { .. }
-                | Effect::PossibleAliasing { .. }
-                | Effect::Drop { .. }
-                | Effect::PinnedReference { .. }
-        )
-    });
-    let inner_non_opacity_pure = inner.effects.effects.iter().all(|e| {
-        matches!(
-            e,
-            Effect::OpaqueLoop { .. }
-                | Effect::EarlyReturn { .. }
-                | Effect::ClosureCapture { .. }
-                | Effect::UnresolvedCall { .. }
-                | Effect::PossibleAliasing { .. }
-                | Effect::Drop { .. }
-                | Effect::PinnedReference { .. }
-        )
-    });
-    if !outer_non_opacity_pure || !inner_non_opacity_pure {
-        return Ok(None);
-    }
-
-    if formal_idx >= outer.formals.len() {
-        return Ok(None);
-    }
-
-    // Phase 3: perform the actual composition (same logic as
-    // compose_function_contracts, duplicated here to avoid the
-    // is_pure() gate in that function).
-    let inner_result_name = inner.result_var_name();
-    let inner_post_renamed = crate::wp::substitute_in_formula(
-        inner.post.clone(),
-        "result",
-        &IrTerm::Var {
-            name: inner_result_name.clone(),
-        },
-    );
-    let inner_value = match find_result_equation(&inner_post_renamed, &inner_result_name) {
-        Some(t) => t,
-        None => return Ok(None),
-    };
-    let outer_formal = &outer.formals[formal_idx];
-    let outer_pre_substituted =
-        crate::wp::substitute_in_formula(outer.pre.clone(), outer_formal, &inner_value);
-    let outer_post_substituted =
-        crate::wp::substitute_in_formula(outer.post.clone(), outer_formal, &inner_value);
-    let pre = IrFormula::And {
-        operands: vec![
-            inner.pre.clone(),
-            IrFormula::Implies {
-                operands: vec![inner_post_renamed, outer_pre_substituted],
-            },
-        ],
-    };
-    let post = outer_post_substituted;
-    let value = Value::object([
-        ("schemaVersion", Value::string("1")),
-        ("kind", Value::string("composed-function-contract")),
-        (
-            "components",
-            Value::array(vec![
-                Value::string(outer.cid.clone()),
-                Value::string(inner.cid.clone()),
-            ]),
-        ),
-        ("formalIdx", Value::integer(formal_idx as i64)),
-        ("pre", formula_to_canonical(&pre)),
-        ("post", formula_to_canonical(&post)),
-    ]);
-    let canonical_bytes = jcs_bytes_of_value(&value);
-    let cid = cid_of_value(&value);
-    Ok(Some(ComposedFunctionContract {
-        component_cids: vec![outer.cid.clone(), inner.cid.clone()],
-        formal_idx,
-        pre,
-        post,
-        canonical_bytes,
-        cid,
-    }))
-}
-
-/// Compose with the inner being an already-composed contract. Used
-/// during chain folding so each step keeps composing without losing
-/// the previous composition's CID.
-pub fn compose_with_composed(
-    outer: &FunctionContractMemento,
-    inner: &ComposedFunctionContract,
-    formal_idx: usize,
-) -> Option<ComposedFunctionContract> {
-    if !outer.is_pure() {
-        return None;
-    }
-    if formal_idx >= outer.formals.len() {
-        return None;
-    }
-
-    // The inner ComposedFunctionContract's post still has the result-
-    // equation form (because outer's post was substituted into it
-    // during the prior compose step). Extract it directly without
-    // renaming — the inner's CID-namespaced result name is preserved.
-    let inner_value = find_result_equation(&inner.post, "result").or_else(|| {
-        // Fallback: scan for any result__-prefixed equation.
-        find_namespaced_result(&inner.post)
-    })?;
-
-    let outer_formal = &outer.formals[formal_idx];
-    let outer_pre_substituted =
-        substitute_in_formula(outer.pre.clone(), outer_formal, &inner_value);
-    let outer_post_substituted =
-        substitute_in_formula(outer.post.clone(), outer_formal, &inner_value);
-
-    let pre = IrFormula::And {
-        operands: vec![
-            inner.pre.clone(),
-            IrFormula::Implies {
-                operands: vec![inner.post.clone(), outer_pre_substituted],
-            },
-        ],
-    };
-    let post = outer_post_substituted;
-
-    let mut components = vec![outer.cid.clone()];
-    components.extend(inner.component_cids.iter().cloned());
-    let component_values: Vec<Arc<Value>> = components
-        .iter()
-        .map(|c| Value::string(c.clone()))
-        .collect();
-    let value = Value::object([
-        ("schemaVersion", Value::string("1")),
-        ("kind", Value::string("composed-function-contract")),
-        ("components", Value::array(component_values)),
-        ("formalIdx", Value::integer(formal_idx as i64)),
-        ("pre", formula_to_canonical(&pre)),
-        ("post", formula_to_canonical(&post)),
-    ]);
-    let canonical_bytes = jcs_bytes_of_value(&value);
-    let cid = cid_of_value(&value);
-
-    Some(ComposedFunctionContract {
-        component_cids: components,
-        formal_idx,
-        pre,
-        post,
-        canonical_bytes,
-        cid,
-    })
-}
-
-/// One step in an N-deep chain composition. The contract receives the
-/// previous step's result at `formal_idx`. The first step's formal_idx
-/// is unused (it's the chain's source).
-#[derive(Debug, Clone, Copy)]
-pub struct ChainStep<'a> {
-    pub contract: &'a FunctionContractMemento,
-    pub formal_idx: usize,
-}
-
-/// Compose a chain of pure function contracts left-to-right. Each
-/// step's contract receives the previous step's result at its
-/// `formal_idx`-th formal. Returns None if any contract is impure or
-/// the chain is shorter than 2 steps.
-///
-/// The chain's overall CID is derivable from its component CIDs in
-/// order — re-composing the same chain produces the same CID
-/// byte-for-byte.
-pub fn compose_chain_contracts(steps: &[ChainStep<'_>]) -> Option<ComposedFunctionContract> {
-    if steps.len() < 2 {
-        return None;
-    }
-    let mut acc =
-        compose_function_contracts(steps[1].contract, steps[0].contract, steps[1].formal_idx)?;
-    for step in &steps[2..] {
-        acc = compose_with_composed(step.contract, &acc, step.formal_idx)?;
-    }
-    Some(acc)
-}
-
-fn find_namespaced_result(formula: &IrFormula) -> Option<IrTerm> {
-    match formula {
-        IrFormula::Atomic { name, args } if name == "=" && args.len() == 2 => {
-            for (var_arg, value_arg) in [(&args[0], &args[1]), (&args[1], &args[0])] {
-                if let IrTerm::Var { name: n } = var_arg {
-                    if n.starts_with("result__") {
-                        return Some(value_arg.clone());
-                    }
-                }
-            }
-            None
-        }
-        IrFormula::And { operands } => operands.iter().find_map(find_namespaced_result),
-        _ => None,
     }
 }
 
@@ -1212,8 +143,6 @@ fn scan_expr_for_effects(expr: &Expr, set: &mut EffectSet) {
         }
         Expr::Macro(m) => scan_macro_for_effects(&m.mac, set),
         Expr::Call(c) => {
-            // Direct callsite — we don't know the callee's effects without
-            // a substrate lookup, so mark unresolved.
             if let Expr::Path(p) = c.func.as_ref() {
                 if let Some(seg) = p.path.segments.last() {
                     let name = seg.ident.to_string();
@@ -1278,7 +207,6 @@ fn scan_expr_for_effects(expr: &Expr, set: &mut EffectSet) {
             }
         }
         Expr::Assign(a) => {
-            // x = expr → writes x.
             if let Expr::Path(p) = a.left.as_ref() {
                 if let Some(seg) = p.path.segments.last() {
                     set.add(Effect::Writes {
@@ -1316,7 +244,7 @@ fn scan_expr_for_effects(expr: &Expr, set: &mut EffectSet) {
                 scan_expr_for_effects(e, set);
             }
         }
-        // Pure expression shapes — no effects to add.
+        // Pure expression shapes; no effects to add.
         Expr::Lit(_) | Expr::Path(_) | Expr::Closure(_) => {}
         _ => {}
     }
@@ -1330,14 +258,10 @@ fn scan_macro_for_effects(mac: &syn::Macro, set: &mut EffectSet) {
         .map(|s| s.ident.to_string())
         .unwrap_or_default();
     match name.as_str() {
-        // Panic-shaped macros.
         "panic" | "unreachable" | "todo" | "unimplemented" => set.add(Effect::Panics),
-        // IO macros.
         "println" | "print" | "eprintln" | "eprint" | "dbg" => set.add(Effect::Io),
-        // Pure (compile-time) macros.
         "assert" | "debug_assert" | "assert_eq" | "assert_ne" => {}
         "vec" | "format" | "concat" | "stringify" => {}
-        // Unknown — conservative mark unresolved.
         _ => set.add(Effect::UnresolvedCall {
             name: format!("{}!", name),
         }),
@@ -1396,13 +320,12 @@ fn is_known_pure_method(name: &str) -> bool {
 }
 
 fn is_known_pure_call(name: &str) -> bool {
-    // Free functions known pure. Conservative; grow over time.
     matches!(name, "min" | "max" | "abs")
 }
 
 // ---- helpers ----
 
-fn extract_formals(item_fn: &ItemFn) -> (Vec<String>, Vec<Sort>) {
+fn extract_formals(item_fn: &ItemFn) -> (Vec<String>, Vec<provekit_ir_types::Sort>) {
     let mut names = Vec::new();
     let mut sorts = Vec::new();
     for input in &item_fn.sig.inputs {
@@ -1418,48 +341,18 @@ fn extract_formals(item_fn: &ItemFn) -> (Vec<String>, Vec<Sort>) {
     (names, sorts)
 }
 
-fn extract_return_sort(item_fn: &ItemFn) -> Sort {
+fn extract_return_sort(item_fn: &ItemFn) -> provekit_ir_types::Sort {
     match &item_fn.sig.output {
-        syn::ReturnType::Default => Sort::Primitive {
+        syn::ReturnType::Default => provekit_ir_types::Sort::Primitive {
             name: "Unit".to_string(),
         },
         syn::ReturnType::Type(_, ty) => infer_sort(ty),
     }
 }
 
-fn infer_sort(ty: &syn::Type) -> Sort {
+fn infer_sort(ty: &syn::Type) -> provekit_ir_types::Sort {
     crate::sort_translate::syn_type_to_sort(ty)
 }
-
-/// Find a top-level `<var> = <expr>` equation in a formula and return
-/// the expr term. Used to extract a function's result-value from its
-/// post for composition.
-fn find_result_equation(formula: &IrFormula, var_name: &str) -> Option<IrTerm> {
-    match formula {
-        IrFormula::Atomic { name, args } if name == "=" && args.len() == 2 => {
-            // Recognize either `var = expr` or `expr = var`.
-            if let IrTerm::Var { name: n } = &args[0] {
-                if n == var_name {
-                    return Some(args[1].clone());
-                }
-            }
-            if let IrTerm::Var { name: n } = &args[1] {
-                if n == var_name {
-                    return Some(args[0].clone());
-                }
-            }
-            None
-        }
-        IrFormula::And { operands } => operands
-            .iter()
-            .find_map(|f| find_result_equation(f, var_name)),
-        _ => None,
-    }
-}
-
-// Suppress unused-import warning in dev builds.
-#[allow(dead_code)]
-fn _unused_wp_path(_w: &Wp) {}
 
 #[cfg(test)]
 mod tests {
@@ -1583,14 +476,10 @@ mod tests {
 
     #[test]
     fn compose_two_pure_contracts_succeeds() {
-        // outer: f(x) = x * 2, post says result = x * 2
-        // inner: g(y) = y + 1, post says result = y + 1
-        // composition f(g(y)): pre = ∅, post = result = (y + 1) * 2
         let f = build_function_contract(&parse_fn(r#"fn f(x: u32) -> u32 { x * 2 }"#), None);
         let g = build_function_contract(&parse_fn(r#"fn g(y: u32) -> u32 { y + 1 }"#), None);
         let composed = compose_function_contracts(&f, &g, 0).expect("compose succeeds");
         assert!(composed.cid.starts_with("blake3-512:"));
-        // Re-composing yields same CID.
         let composed2 = compose_function_contracts(&f, &g, 0).unwrap();
         assert_eq!(composed.cid, composed2.cid);
     }
@@ -1625,7 +514,6 @@ mod tests {
         let f = build_function_contract(&parse_fn(r#"fn f(x: u32) -> u32 { x * 2 }"#), None);
         let name = f.result_var_name();
         assert!(name.starts_with("result__"));
-        // Different functions get different namespaces.
         let g = build_function_contract(&parse_fn(r#"fn g(y: u32) -> u32 { y * 3 }"#), None);
         assert_ne!(f.result_var_name(), g.result_var_name());
     }
@@ -1647,7 +535,6 @@ mod tests {
         assert_eq!(a.cid, b.cid);
     }
 
-    // Suppress unused-import warnings.
     #[test]
     fn _unused_helpers() {
         let _ = atomic_ge(var("x"), const_int(1));
@@ -1655,9 +542,6 @@ mod tests {
 
     // ---- Bug #384 A.1: sort-collapse regression tests ----
 
-    /// fn f(x: u32) and fn f(x: bool) must produce DISTINCT contracts
-    /// (distinct formal_sorts → distinct content_cid). This was the
-    /// primary false-collision from the old token-string infer_sort.
     #[test]
     fn u32_and_bool_formals_produce_distinct_cids() {
         let f_u32 = build_function_contract(&parse_fn(r#"fn f(x: u32) {}"#), None);
@@ -1669,10 +553,6 @@ mod tests {
         assert_ne!(f_u32.formal_sorts, f_bool.formal_sorts);
     }
 
-    /// Lifetime annotation on a reference must NOT change the formal sort
-    /// or the contract CID. This was the false-split from whitespace
-    /// tokenisation: &'a str tokenised as "& 'a str" (with spaces)
-    /// fell to the catch-all, while &str matched the explicit arm.
     #[test]
     fn ref_lifetime_annotation_does_not_change_cid() {
         let with_lt = build_function_contract(&parse_fn(r#"fn f<'a>(s: &'a str) {}"#), None);
@@ -1681,11 +561,8 @@ mod tests {
             with_lt.formal_sorts, without_lt.formal_sorts,
             "&'a str and &str must produce the same formal sort"
         );
-        // CIDs will differ because fn names include the lifetime parameter
-        // in the AST (`fn f<'a>` vs `fn f`), but formal_sorts must agree.
     }
 
-    /// Vec<u32> and a user struct produce distinct formal sorts.
     #[test]
     fn vec_and_user_struct_formals_are_distinct() {
         let f_vec = build_function_contract(&parse_fn(r#"fn f(x: Vec<u32>) {}"#), None);
@@ -1699,7 +576,6 @@ mod tests {
 
     // ---- Issue #384 B.5: opacity discharge tests ----
 
-    /// Mock pool for testing: pre-seeded with specific loop_cid values.
     struct MockPool {
         loop_cids: Vec<String>,
         try_cids: Vec<String>,
@@ -1751,10 +627,10 @@ mod tests {
             self.body_fn_cids.iter().any(|c| c == body_fn_cid)
         }
         fn has_drop_contract(&self, _: &str) -> bool {
-            false // no drop contracts in mock pool
+            false
         }
         fn has_aliasing_memento(&self, _a: &str, _b: &str) -> bool {
-            false // no aliasing mementos in mock pool
+            false
         }
         fn lookup_pin_invariant(
             &self,
@@ -1773,8 +649,6 @@ mod tests {
         }
     }
 
-    /// Build a FunctionContractMemento from a `fn f(x: u32) -> u32` shell
-    /// with a manually injected effect. Used by opacity tests below.
     fn contract_with_effects(name: &str, effects: Vec<Effect>) -> FunctionContractMemento {
         let mut c = build_function_contract(
             &parse_fn(&format!("fn {}(x: u32) -> u32 {{ x }}", name)),
@@ -1783,18 +657,15 @@ mod tests {
         for e in effects {
             c.effects.add(e);
         }
-        // Recompute cid so it reflects the new effects set.
-        let val = crate::contract::build_memento_value(&c);
-        c.canonical_bytes = crate::canonical::jcs_bytes_of_value(&val);
-        c.cid = crate::canonical::cid_of_value(&val);
+        let val = build_memento_value(&c);
+        c.canonical_bytes = jcs_bytes_of_value(&val);
+        c.cid = cid_of_value(&val);
         c
     }
 
-    /// A contract with OpaqueLoop + no LoopInvariantMemento in pool →
-    /// compose_function_contracts_checked returns Err.
     #[test]
     fn opaque_loop_without_memento_blocks_composition() {
-        let loop_cid = "blake3-512:aabb".repeat(8); // fake stable cid
+        let loop_cid = "blake3-512:aabb".repeat(8);
         let outer = contract_with_effects(
             "outer",
             vec![Effect::OpaqueLoop {
@@ -1803,7 +674,7 @@ mod tests {
         );
         let inner =
             build_function_contract(&parse_fn(r#"fn inner(y: u32) -> u32 { y + 1 }"#), None);
-        let pool = MockPool::empty(); // no loop invariant
+        let pool = MockPool::empty();
         let result = compose_function_contracts_checked(&outer, &inner, 0, &pool);
         assert!(
             matches!(result, Err(OpacityError::LoopNotDischarged { .. })),
@@ -1812,8 +683,6 @@ mod tests {
         );
     }
 
-    /// Same contract with OpaqueLoop + matching LoopInvariantMemento in pool →
-    /// compose_function_contracts_checked succeeds (returns Ok(Some(...))).
     #[test]
     fn opaque_loop_with_memento_allows_composition() {
         let loop_cid = "blake3-512:aabb".repeat(8);
@@ -1827,8 +696,6 @@ mod tests {
             build_function_contract(&parse_fn(r#"fn inner(y: u32) -> u32 { y + 1 }"#), None);
         let pool = MockPool::empty().with_loop(&loop_cid);
         let result = compose_function_contracts_checked(&outer, &inner, 0, &pool);
-        // outer has only an OpaqueLoop (now discharged) and inner is pure →
-        // composition should succeed.
         assert!(
             matches!(result, Ok(Some(_))),
             "expected Ok(Some(_)) after discharge, got {:?}",
@@ -1836,7 +703,6 @@ mod tests {
         );
     }
 
-    /// EarlyReturn + no TryBranchMemento → Err(EarlyReturnNotDischarged).
     #[test]
     fn early_return_without_memento_blocks_composition() {
         let try_cid = "blake3-512:ccdd".repeat(8);
@@ -1856,7 +722,6 @@ mod tests {
         );
     }
 
-    /// EarlyReturn + matching TryBranchMemento → Ok(Some(_)).
     #[test]
     fn early_return_with_memento_allows_composition() {
         let try_cid = "blake3-512:ccdd".repeat(8);
@@ -1876,7 +741,6 @@ mod tests {
         );
     }
 
-    /// ClosureCapture + no ClosureBindingMemento → Err(ClosureCaptureNotDischarged).
     #[test]
     fn closure_capture_without_memento_blocks_composition() {
         let body_fn_cid = "blake3-512:eeff".repeat(8);
@@ -1900,7 +764,6 @@ mod tests {
         );
     }
 
-    /// ClosureCapture + matching ClosureBindingMemento → Ok(Some(_)).
     #[test]
     fn closure_capture_with_memento_allows_composition() {
         let body_fn_cid = "blake3-512:eeff".repeat(8);
@@ -1921,7 +784,6 @@ mod tests {
         );
     }
 
-    /// UnresolvedCall always blocks composition regardless of pool contents.
     #[test]
     fn unresolved_call_always_blocks_composition() {
         let outer = contract_with_effects(
@@ -1943,12 +805,9 @@ mod tests {
         );
     }
 
-    /// Non-opacity impure effects (Io) still block composition even
-    /// after all opacity effects are discharged: returns Ok(None).
     #[test]
     fn non_opacity_effects_still_block_after_opacity_discharge() {
         let loop_cid = "blake3-512:1122".repeat(8);
-        // Contract has both OpaqueLoop (dischargeable) and Io (not dischargeable).
         let outer = contract_with_effects(
             "outer",
             vec![
@@ -1961,7 +820,6 @@ mod tests {
         let inner = build_function_contract(&parse_fn(r#"fn inner(y: u32) -> u32 { y }"#), None);
         let pool = MockPool::empty().with_loop(&loop_cid);
         let result = compose_function_contracts_checked(&outer, &inner, 0, &pool);
-        // OpaqueLoop is discharged → no OpacityError. But Io blocks → Ok(None).
         assert!(
             matches!(result, Ok(None)),
             "expected Ok(None) when non-opacity effect blocks, got {:?}",
@@ -1969,7 +827,6 @@ mod tests {
         );
     }
 
-    /// check_opacity_effects returns Ok(()) for a pure contract.
     #[test]
     fn check_opacity_pure_contract_is_ok() {
         let f = build_function_contract(&parse_fn(r#"fn f(x: u32) -> u32 { x * 2 }"#), None);
@@ -1979,8 +836,6 @@ mod tests {
 
     // ---- Issue #395: PinInvariantMemento discharge tests ----
 
-    /// A contract with PinnedReference + no PinInvariantMemento in pool ->
-    /// compose_function_contracts_checked returns Err(PinInvariantNotDischarged).
     #[test]
     fn pinned_reference_without_memento_blocks_composition() {
         let outer = contract_with_effects(
@@ -2000,8 +855,6 @@ mod tests {
         );
     }
 
-    /// Same contract with PinnedReference + matching PinInvariantMemento in pool ->
-    /// compose_function_contracts_checked succeeds.
     #[test]
     fn pinned_reference_with_memento_succeeds() {
         let outer = contract_with_effects(
@@ -2023,8 +876,6 @@ mod tests {
         }
     }
 
-    /// Pool has a PinInvariantMemento for a different target -> composed_function_contracts_checked
-    /// returns Err(PinInvariantNotDischarged).
     #[test]
     fn pinned_reference_with_wrong_target_memento_blocks() {
         let outer = contract_with_effects(
@@ -2044,8 +895,6 @@ mod tests {
         );
     }
 
-    /// Pool has a PinInvariantMemento with empty invariant -> still blocks
-    /// because non-emptiness is required by the spec (§3.2).
     #[test]
     fn pinned_reference_with_empty_invariant_blocks() {
         let outer = contract_with_effects(
